Specify platform for Docker services
Currently when we don't specify platform for services in Docker, it defaults
building images using the local architecture. This poses a problem with
M1 Macs, as our images don't all build successfully using arm.

To get around this for now, I'm conservatively specifying the architecture
for services on architecture we know works, which will then be built using
that emulator on M1 Macs. The one exception to this is the check-api repo,
because there is a bug preventing the image from being built with the emulator.
The issue is described here: https://github.com/docker/for-mac/issues/5321

Fixes CHECK-2147