add -e to the whole script and set +e when the diff is run
-e checks if a value other than 0 is returned, and exists the script.
So it's loking for errors. I guess that is usefull to us since we can
get errors during the build.

However when we have a diff between files it returns 1, meaning the
script quits once that happens. To keep the -e flag I'm only 'turning
it off' once we are running the diffs, by running set +e before that.