3082: Handle changes to configuration files in local development environment (#77)
* add diff-test

We want to make the user aware of any new changes to the config file.
Currently I'm testing a few things.

* add conditional to be clearer when there are no config files

* add checking the config to git-update

I changed the configs to be an array, that was missing before, and also
tested implementing  it inside the git-update script

* change it to make a copy and compare it to pulled file

I was thinking, git diff won't really work here.

So maybe if we copy the local files, update the repo and compare those
using diff (not the same as git diff) it might work better.

It's super messy, but I think it's going somewhere

* make subdirectories for services with multiple configs

currently it's only saving the last config from check-bots, and running
all of them against the last one

* refactor: remove deprecated methods and add double quotes

* find all files with .example

* rethink how this could be written

Writing everything inside git submodule foreach can
make it a little harder to read. It's also not needed.

I was thinking a lot in terms of git instead of thinking about
how we only need to go through the directories and files.

* merge the two scripts

For a moment I thought I could maybe just call the diff script at the end,
but I forgot that I need to store the local changes before I pull. So I need
to rethink the best way for these two to work together

* add -e to the whole script and set +e when the diff is run

-e checks if a value other than 0 is returned, and exists the script.
So it's looking for errors. I guess that is useful to us since we can
get errors during the build.

However when we have a diff between files it returns 1, meaning the
script quits once that happens. To keep the -e flag I'm only 'turning
it off' once we are running the diffs, by running set +e before that.

* delete test files

* update to  use variable instead of function

when I was first testing things I used a function to build the path and
get it again later, but a variable is enough

* update diff to have a shorter output

* update readme to reflect changes to git-update

* update to deal with ShellCheck warning

SC2207 – ShellCheck Wiki
Prefer mapfile or read -a to split command output (or quote to avoid splitting).

There seems to be some issues building arrays like this
configs=($(find . -name '*.example' -not -path '*/temp/*'))
references:
http://mywiki.wooledge.org/BashPitfalls#for_f_in_.24.28ls_.2A.mp3.29
https://unix.stackexchange.com/questions/715453/simple-loop-of-files-file-name-too-long