4755 – Add otel-collector to docker compose (#95)
We use Honeycomb+otel for our observability, unfortunately otel does not provide a metrics client for Ruby.
In order to send metrics to Honeycomb we instrumented Pender with the prometheus ruby-client, and used otel-collector to scrape the Prometheus endpoint and send those metrics to Honeycomb.

I added the bare minimum configuration for it to work on Pender's side (PR here: meedan/pender#456). There are features that might be interesting, but I think we could add those as we add the metrics and understand what makes sense for us. Depending on what features we decide to use, we might need to update this file.

Resources
- https://docs.honeycomb.io/integrations/metrics/prometheus/
- https://opentelemetry.io/docs/collector/installation/

References: 4755, 4396
PR: 95