6473 – add otel config (#108)
We are using the otel-collector mainly to be able to collect custom metrics from our Rails services. But it is a service that can be used with any of Check's services.

Because it is relevant to all of our services we don't want to keep it in individual repos, but keep it in the main Check one.

This is only relevant to local development.

References: CV2-6473, CV2-4396
Relates to: meedan/pender#586