Disable spring
I think there is some weirdness because of our mount bind. When running
from the pender repo it works well, but when running from the check repo
we hit: Error connecting to Spring server.

I tried setting the SPRING_SOCKET and SPRING_TMP_PATH, but couldn't
get it to work.