Disable spring (#111)
Context
When running Pender from the main repo, and the running tests, tests would hang. What I mean by hang:
    - If you run one or multiple tests,
    - the run finishes,
    - when you try to run another tests, it just hangs there and doesn’t print anything or return any information

Issue
This seems to be caused by a failure to connect to the Spring server.  
I think this might be a combination of how spring, the sockets it uses,
our mount bind, the way we have this set up so we have can run from
both check and Pender (so we have two docker-composes), and maybe
docker-mac issues.

Fixing this
I tried setting the SPRING_SOCKET and SPRING_TMP_PATH, but couldn't
get it to work. Since trying to fix this is taking too long (it can take too long
between builds), and it isn't a big deal to just remove this, I disabled Spring
from the main Check repo. It is still enabled when running tests from the 
Pender repo.

References: CV2-6431
Relates to: https://github.com/meedan/pender/pull/593