Add routing connector to OTEL config
So we can route a specific service prometheus metrics
to the specific service honeycomb dataset.

Notes:
1. We don't need to send the dataset when sending traces,
but we still do when sending metrics.
2. While we can aliases most otel-collector components, e.g.:
otlp/pender_metrics. The prometheus receiver does not support that,
so we need the router connector to send to the right pipeline.
(There is a target-allocation feature in the prometheus receiver,
but it is specific to kubernetes.)